upate slack config, use full method name
