+++ backend/api/config/service/auditsink/slack/v1/slack.pb.validate.go
@@ -134,3 +134,163 @@ var _ interface {
 	Cause() error
 	ErrorName() string
 } = SlackConfigValidationError{}
+
+// Validate checks the field values on CustomSlackMessage with the rules
+// defined in the proto definition for this message. If any rules are
+// violated, an error is returned.
+func (m *CustomSlackMessage) Validate() error {
+	if m == nil {
+		return nil
+	}
+
+	if len(m.GetFullMethod()) < 1 {
+		return CustomSlackMessageValidationError{
+			field:  "FullMethod",
+			reason: "value length must be at least 1 bytes",
+		}
+	}
+
+	if len(m.GetMessage()) < 1 {
+		return CustomSlackMessageValidationError{
+			field:  "Message",
+			reason: "value length must be at least 1 bytes",
+		}
+	}
+
+	return nil
+}
+
+// CustomSlackMessageValidationError is the validation error returned by
+// CustomSlackMessage.Validate if the designated constraints aren't met.
+type CustomSlackMessageValidationError struct {
+	field  string
+	reason string
+	cause  error
+	key    bool
+}
+
+// Field function returns field value.
+func (e CustomSlackMessageValidationError) Field() string { return e.field }
+
+// Reason function returns reason value.
+func (e CustomSlackMessageValidationError) Reason() string { return e.reason }
+
+// Cause function returns cause value.
+func (e CustomSlackMessageValidationError) Cause() error { return e.cause }
+
+// Key function returns key value.
+func (e CustomSlackMessageValidationError) Key() bool { return e.key }
+
+// ErrorName returns error name.
+func (e CustomSlackMessageValidationError) ErrorName() string {
+	return "CustomSlackMessageValidationError"
+}
+
+// Error satisfies the builtin error interface
+func (e CustomSlackMessageValidationError) Error() string {
+	cause := ""
+	if e.cause != nil {
+		cause = fmt.Sprintf(" | caused by: %v", e.cause)
+	}
+
+	key := ""
+	if e.key {
+		key = "key for "
+	}
+
+	return fmt.Sprintf(
+		"invalid %sCustomSlackMessage.%s: %s%s",
+		key,
+		e.field,
+		e.reason,
+		cause)
+}
+
+var _ error = CustomSlackMessageValidationError{}
+
+var _ interface {
+	Field() string
+	Reason() string
+	Key() bool
+	Cause() error
+	ErrorName() string
+} = CustomSlackMessageValidationError{}
+
+// Validate checks the field values on Override with the rules defined in the
+// proto definition for this message. If any rules are violated, an error is returned.
+func (m *Override) Validate() error {
+	if m == nil {
+		return nil
+	}
+
+	for idx, item := range m.GetCustomSlackMessages() {
+		_, _ = idx, item
+
+		if v, ok := interface{}(item).(interface{ Validate() error }); ok {
+			if err := v.Validate(); err != nil {
+				return OverrideValidationError{
+					field:  fmt.Sprintf("CustomSlackMessages[%v]", idx),
+					reason: "embedded message failed validation",
+					cause:  err,
+				}
+			}
+		}
+
+	}
+
+	return nil
+}
+
+// OverrideValidationError is the validation error returned by
+// Override.Validate if the designated constraints aren't met.
+type OverrideValidationError struct {
+	field  string
+	reason string
+	cause  error
+	key    bool
+}
+
+// Field function returns field value.
+func (e OverrideValidationError) Field() string { return e.field }
+
+// Reason function returns reason value.
+func (e OverrideValidationError) Reason() string { return e.reason }
+
+// Cause function returns cause value.
+func (e OverrideValidationError) Cause() error { return e.cause }
+
+// Key function returns key value.
+func (e OverrideValidationError) Key() bool { return e.key }
+
+// ErrorName returns error name.
+func (e OverrideValidationError) ErrorName() string { return "OverrideValidationError" }
+
+// Error satisfies the builtin error interface
+func (e OverrideValidationError) Error() string {
+	cause := ""
+	if e.cause != nil {
+		cause = fmt.Sprintf(" | caused by: %v", e.cause)
+	}
+
+	key := ""
+	if e.key {
+		key = "key for "
+	}
+
+	return fmt.Sprintf(
+		"invalid %sOverride.%s: %s%s",
+		key,
+		e.field,
+		e.reason,
+		cause)
+}
+
+var _ error = OverrideValidationError{}
+
+var _ interface {
+	Field() string
+	Reason() string
+	Key() bool
+	Cause() error
+	ErrorName() string
+} = OverrideValidationError{}
