+++ frontend/api/src/index.d.ts
@@ -6023,7 +6023,7 @@ export namespace clutch {
                             filter?: (clutch.config.service.audit.v1.IFilter|null);
 
                             /** SlackConfig override */
-                            override?: (clutch.config.service.audit.v1.IOverride|null);
+                            override?: (clutch.config.service.auditsink.slack.v1.IOverride|null);
                         }
 
                         /** Represents a SlackConfig. */
@@ -6045,7 +6045,7 @@ export namespace clutch {
                             public filter?: (clutch.config.service.audit.v1.IFilter|null);
 
                             /** SlackConfig override. */
-                            public override?: (clutch.config.service.audit.v1.IOverride|null);
+                            public override?: (clutch.config.service.auditsink.slack.v1.IOverride|null);
 
                             /**
                              * Verifies a SlackConfig message.
@@ -6071,6 +6071,108 @@ export namespace clutch {
 
                             /**
                              * Converts this SlackConfig to JSON.
+                             * @returns JSON object
+                             */
+                            public toJSON(): { [k: string]: any };
+                        }
+
+                        /** Properties of a CustomSlackMessage. */
+                        interface ICustomSlackMessage {
+
+                            /** CustomSlackMessage fullMethod */
+                            fullMethod?: (string|null);
+
+                            /** CustomSlackMessage message */
+                            message?: (string|null);
+                        }
+
+                        /** Represents a CustomSlackMessage. */
+                        class CustomSlackMessage implements ICustomSlackMessage {
+
+                            /**
+                             * Constructs a new CustomSlackMessage.
+                             * @param [properties] Properties to set
+                             */
+                            constructor(properties?: clutch.config.service.auditsink.slack.v1.ICustomSlackMessage);
+
+                            /** CustomSlackMessage fullMethod. */
+                            public fullMethod: string;
+
+                            /** CustomSlackMessage message. */
+                            public message: string;
+
+                            /**
+                             * Verifies a CustomSlackMessage message.
+                             * @param message Plain object to verify
+                             * @returns `null` if valid, otherwise the reason why it is not
+                             */
+                            public static verify(message: { [k: string]: any }): (string|null);
+
+                            /**
+                             * Creates a CustomSlackMessage message from a plain object. Also converts values to their respective internal types.
+                             * @param object Plain object
+                             * @returns CustomSlackMessage
+                             */
+                            public static fromObject(object: { [k: string]: any }): clutch.config.service.auditsink.slack.v1.CustomSlackMessage;
+
+                            /**
+                             * Creates a plain object from a CustomSlackMessage message. Also converts values to other types if specified.
+                             * @param message CustomSlackMessage
+                             * @param [options] Conversion options
+                             * @returns Plain object
+                             */
+                            public static toObject(message: clutch.config.service.auditsink.slack.v1.CustomSlackMessage, options?: $protobuf.IConversionOptions): { [k: string]: any };
+
+                            /**
+                             * Converts this CustomSlackMessage to JSON.
+                             * @returns JSON object
+                             */
+                            public toJSON(): { [k: string]: any };
+                        }
+
+                        /** Properties of an Override. */
+                        interface IOverride {
+
+                            /** Override customSlackMessages */
+                            customSlackMessages?: (clutch.config.service.auditsink.slack.v1.ICustomSlackMessage[]|null);
+                        }
+
+                        /** Represents an Override. */
+                        class Override implements IOverride {
+
+                            /**
+                             * Constructs a new Override.
+                             * @param [properties] Properties to set
+                             */
+                            constructor(properties?: clutch.config.service.auditsink.slack.v1.IOverride);
+
+                            /** Override customSlackMessages. */
+                            public customSlackMessages: clutch.config.service.auditsink.slack.v1.ICustomSlackMessage[];
+
+                            /**
+                             * Verifies an Override message.
+                             * @param message Plain object to verify
+                             * @returns `null` if valid, otherwise the reason why it is not
+                             */
+                            public static verify(message: { [k: string]: any }): (string|null);
+
+                            /**
+                             * Creates an Override message from a plain object. Also converts values to their respective internal types.
+                             * @param object Plain object
+                             * @returns Override
+                             */
+                            public static fromObject(object: { [k: string]: any }): clutch.config.service.auditsink.slack.v1.Override;
+
+                            /**
+                             * Creates a plain object from an Override message. Also converts values to other types if specified.
+                             * @param message Override
+                             * @param [options] Conversion options
+                             * @returns Plain object
+                             */
+                            public static toObject(message: clutch.config.service.auditsink.slack.v1.Override, options?: $protobuf.IConversionOptions): { [k: string]: any };
+
+                            /**
+                             * Converts this Override to JSON.
                              * @returns JSON object
                              */
                             public toJSON(): { [k: string]: any };
@@ -6202,108 +6304,6 @@ export namespace clutch {
 
                         /**
                          * Converts this Filter to JSON.
-                         * @returns JSON object
-                         */
-                        public toJSON(): { [k: string]: any };
-                    }
-
-                    /** Properties of a CustomSlackMessage. */
-                    interface ICustomSlackMessage {
-
-                        /** CustomSlackMessage method */
-                        method?: (string|null);
-
-                        /** CustomSlackMessage message */
-                        message?: (string|null);
-                    }
-
-                    /** Represents a CustomSlackMessage. */
-                    class CustomSlackMessage implements ICustomSlackMessage {
-
-                        /**
-                         * Constructs a new CustomSlackMessage.
-                         * @param [properties] Properties to set
-                         */
-                        constructor(properties?: clutch.config.service.audit.v1.ICustomSlackMessage);
-
-                        /** CustomSlackMessage method. */
-                        public method: string;
-
-                        /** CustomSlackMessage message. */
-                        public message: string;
-
-                        /**
-                         * Verifies a CustomSlackMessage message.
-                         * @param message Plain object to verify
-                         * @returns `null` if valid, otherwise the reason why it is not
-                         */
-                        public static verify(message: { [k: string]: any }): (string|null);
-
-                        /**
-                         * Creates a CustomSlackMessage message from a plain object. Also converts values to their respective internal types.
-                         * @param object Plain object
-                         * @returns CustomSlackMessage
-                         */
-                        public static fromObject(object: { [k: string]: any }): clutch.config.service.audit.v1.CustomSlackMessage;
-
-                        /**
-                         * Creates a plain object from a CustomSlackMessage message. Also converts values to other types if specified.
-                         * @param message CustomSlackMessage
-                         * @param [options] Conversion options
-                         * @returns Plain object
-                         */
-                        public static toObject(message: clutch.config.service.audit.v1.CustomSlackMessage, options?: $protobuf.IConversionOptions): { [k: string]: any };
-
-                        /**
-                         * Converts this CustomSlackMessage to JSON.
-                         * @returns JSON object
-                         */
-                        public toJSON(): { [k: string]: any };
-                    }
-
-                    /** Properties of an Override. */
-                    interface IOverride {
-
-                        /** Override customSlackMessages */
-                        customSlackMessages?: (clutch.config.service.audit.v1.ICustomSlackMessage[]|null);
-                    }
-
-                    /** Represents an Override. */
-                    class Override implements IOverride {
-
-                        /**
-                         * Constructs a new Override.
-                         * @param [properties] Properties to set
-                         */
-                        constructor(properties?: clutch.config.service.audit.v1.IOverride);
-
-                        /** Override customSlackMessages. */
-                        public customSlackMessages: clutch.config.service.audit.v1.ICustomSlackMessage[];
-
-                        /**
-                         * Verifies an Override message.
-                         * @param message Plain object to verify
-                         * @returns `null` if valid, otherwise the reason why it is not
-                         */
-                        public static verify(message: { [k: string]: any }): (string|null);
-
-                        /**
-                         * Creates an Override message from a plain object. Also converts values to their respective internal types.
-                         * @param object Plain object
-                         * @returns Override
-                         */
-                        public static fromObject(object: { [k: string]: any }): clutch.config.service.audit.v1.Override;
-
-                        /**
-                         * Creates a plain object from an Override message. Also converts values to other types if specified.
-                         * @param message Override
-                         * @param [options] Conversion options
-                         * @returns Plain object
-                         */
-                        public static toObject(message: clutch.config.service.audit.v1.Override, options?: $protobuf.IConversionOptions): { [k: string]: any };
-
-                        /**
-                         * Converts this Override to JSON.
                          * @returns JSON object
                          */
                         public toJSON(): { [k: string]: any };
