+++ frontend/api/src/index.js
@@ -14360,7 +14360,7 @@ export const clutch = $root.clutch = (() => {
                              * @property {string|null} [token] SlackConfig token
                              * @property {string|null} [channel] SlackConfig channel
                              * @property {clutch.config.service.audit.v1.IFilter|null} [filter] SlackConfig filter
-                             * @property {clutch.config.service.audit.v1.IOverride|null} [override] SlackConfig override
+                             * @property {clutch.config.service.auditsink.slack.v1.IOverride|null} [override] SlackConfig override
                              */
 
                             /**
@@ -14404,7 +14404,7 @@ export const clutch = $root.clutch = (() => {
 
                             /**
                              * SlackConfig override.
-                             * @member {clutch.config.service.audit.v1.IOverride|null|undefined} override
+                             * @member {clutch.config.service.auditsink.slack.v1.IOverride|null|undefined} override
                              * @memberof clutch.config.service.auditsink.slack.v1.SlackConfig
                              * @instance
                              */
@@ -14433,7 +14433,7 @@ export const clutch = $root.clutch = (() => {
                                         return "filter." + error;
                                 }
                                 if (message.override != null && message.hasOwnProperty("override")) {
-                                    let error = $root.clutch.config.service.audit.v1.Override.verify(message.override);
+                                    let error = $root.clutch.config.service.auditsink.slack.v1.Override.verify(message.override);
                                     if (error)
                                         return "override." + error;
                                 }
@@ -14464,7 +14464,7 @@ export const clutch = $root.clutch = (() => {
                                 if (object.override != null) {
                                     if (typeof object.override !== "object")
                                         throw TypeError(".clutch.config.service.auditsink.slack.v1.SlackConfig.override: object expected");
-                                    message.override = $root.clutch.config.service.audit.v1.Override.fromObject(object.override);
+                                    message.override = $root.clutch.config.service.auditsink.slack.v1.Override.fromObject(object.override);
                                 }
                                 return message;
                             };
@@ -14495,7 +14495,7 @@ export const clutch = $root.clutch = (() => {
                                 if (message.filter != null && message.hasOwnProperty("filter"))
                                     object.filter = $root.clutch.config.service.audit.v1.Filter.toObject(message.filter, options);
                                 if (message.override != null && message.hasOwnProperty("override"))
-                                    object.override = $root.clutch.config.service.audit.v1.Override.toObject(message.override, options);
+                                    object.override = $root.clutch.config.service.auditsink.slack.v1.Override.toObject(message.override, options);
                                 return object;
                             };
 
@@ -14511,6 +14511,242 @@ export const clutch = $root.clutch = (() => {
                             };
 
                             return SlackConfig;
+                        })();
+
+                        v1.CustomSlackMessage = (function() {
+
+                            /**
+                             * Properties of a CustomSlackMessage.
+                             * @memberof clutch.config.service.auditsink.slack.v1
+                             * @interface ICustomSlackMessage
+                             * @property {string|null} [fullMethod] CustomSlackMessage fullMethod
+                             * @property {string|null} [message] CustomSlackMessage message
+                             */
+
+                            /**
+                             * Constructs a new CustomSlackMessage.
+                             * @memberof clutch.config.service.auditsink.slack.v1
+                             * @classdesc Represents a CustomSlackMessage.
+                             * @implements ICustomSlackMessage
+                             * @constructor
+                             * @param {clutch.config.service.auditsink.slack.v1.ICustomSlackMessage=} [properties] Properties to set
+                             */
+                            function CustomSlackMessage(properties) {
+                                if (properties)
+                                    for (let keys = Object.keys(properties), i = 0; i < keys.length; ++i)
+                                        if (properties[keys[i]] != null)
+                                            this[keys[i]] = properties[keys[i]];
+                            }
+
+                            /**
+                             * CustomSlackMessage fullMethod.
+                             * @member {string} fullMethod
+                             * @memberof clutch.config.service.auditsink.slack.v1.CustomSlackMessage
+                             * @instance
+                             */
+                            CustomSlackMessage.prototype.fullMethod = "";
+
+                            /**
+                             * CustomSlackMessage message.
+                             * @member {string} message
+                             * @memberof clutch.config.service.auditsink.slack.v1.CustomSlackMessage
+                             * @instance
+                             */
+                            CustomSlackMessage.prototype.message = "";
+
+                            /**
+                             * Verifies a CustomSlackMessage message.
+                             * @function verify
+                             * @memberof clutch.config.service.auditsink.slack.v1.CustomSlackMessage
+                             * @static
+                             * @param {Object.<string,*>} message Plain object to verify
+                             * @returns {string|null} `null` if valid, otherwise the reason why it is not
+                             */
+                            CustomSlackMessage.verify = function verify(message) {
+                                if (typeof message !== "object" || message === null)
+                                    return "object expected";
+                                if (message.fullMethod != null && message.hasOwnProperty("fullMethod"))
+                                    if (!$util.isString(message.fullMethod))
+                                        return "fullMethod: string expected";
+                                if (message.message != null && message.hasOwnProperty("message"))
+                                    if (!$util.isString(message.message))
+                                        return "message: string expected";
+                                return null;
+                            };
+
+                            /**
+                             * Creates a CustomSlackMessage message from a plain object. Also converts values to their respective internal types.
+                             * @function fromObject
+                             * @memberof clutch.config.service.auditsink.slack.v1.CustomSlackMessage
+                             * @static
+                             * @param {Object.<string,*>} object Plain object
+                             * @returns {clutch.config.service.auditsink.slack.v1.CustomSlackMessage} CustomSlackMessage
+                             */
+                            CustomSlackMessage.fromObject = function fromObject(object) {
+                                if (object instanceof $root.clutch.config.service.auditsink.slack.v1.CustomSlackMessage)
+                                    return object;
+                                let message = new $root.clutch.config.service.auditsink.slack.v1.CustomSlackMessage();
+                                if (object.fullMethod != null)
+                                    message.fullMethod = String(object.fullMethod);
+                                if (object.message != null)
+                                    message.message = String(object.message);
+                                return message;
+                            };
+
+                            /**
+                             * Creates a plain object from a CustomSlackMessage message. Also converts values to other types if specified.
+                             * @function toObject
+                             * @memberof clutch.config.service.auditsink.slack.v1.CustomSlackMessage
+                             * @static
+                             * @param {clutch.config.service.auditsink.slack.v1.CustomSlackMessage} message CustomSlackMessage
+                             * @param {$protobuf.IConversionOptions} [options] Conversion options
+                             * @returns {Object.<string,*>} Plain object
+                             */
+                            CustomSlackMessage.toObject = function toObject(message, options) {
+                                if (!options)
+                                    options = {};
+                                let object = {};
+                                if (options.defaults) {
+                                    object.fullMethod = "";
+                                    object.message = "";
+                                }
+                                if (message.fullMethod != null && message.hasOwnProperty("fullMethod"))
+                                    object.fullMethod = message.fullMethod;
+                                if (message.message != null && message.hasOwnProperty("message"))
+                                    object.message = message.message;
+                                return object;
+                            };
+
+                            /**
+                             * Converts this CustomSlackMessage to JSON.
+                             * @function toJSON
+                             * @memberof clutch.config.service.auditsink.slack.v1.CustomSlackMessage
+                             * @instance
+                             * @returns {Object.<string,*>} JSON object
+                             */
+                            CustomSlackMessage.prototype.toJSON = function toJSON() {
+                                return this.constructor.toObject(this, $protobuf.util.toJSONOptions);
+                            };
+
+                            return CustomSlackMessage;
+                        })();
+
+                        v1.Override = (function() {
+
+                            /**
+                             * Properties of an Override.
+                             * @memberof clutch.config.service.auditsink.slack.v1
+                             * @interface IOverride
+                             * @property {Array.<clutch.config.service.auditsink.slack.v1.ICustomSlackMessage>|null} [customSlackMessages] Override customSlackMessages
+                             */
+
+                            /**
+                             * Constructs a new Override.
+                             * @memberof clutch.config.service.auditsink.slack.v1
+                             * @classdesc Represents an Override.
+                             * @implements IOverride
+                             * @constructor
+                             * @param {clutch.config.service.auditsink.slack.v1.IOverride=} [properties] Properties to set
+                             */
+                            function Override(properties) {
+                                this.customSlackMessages = [];
+                                if (properties)
+                                    for (let keys = Object.keys(properties), i = 0; i < keys.length; ++i)
+                                        if (properties[keys[i]] != null)
+                                            this[keys[i]] = properties[keys[i]];
+                            }
+
+                            /**
+                             * Override customSlackMessages.
+                             * @member {Array.<clutch.config.service.auditsink.slack.v1.ICustomSlackMessage>} customSlackMessages
+                             * @memberof clutch.config.service.auditsink.slack.v1.Override
+                             * @instance
+                             */
+                            Override.prototype.customSlackMessages = $util.emptyArray;
+
+                            /**
+                             * Verifies an Override message.
+                             * @function verify
+                             * @memberof clutch.config.service.auditsink.slack.v1.Override
+                             * @static
+                             * @param {Object.<string,*>} message Plain object to verify
+                             * @returns {string|null} `null` if valid, otherwise the reason why it is not
+                             */
+                            Override.verify = function verify(message) {
+                                if (typeof message !== "object" || message === null)
+                                    return "object expected";
+                                if (message.customSlackMessages != null && message.hasOwnProperty("customSlackMessages")) {
+                                    if (!Array.isArray(message.customSlackMessages))
+                                        return "customSlackMessages: array expected";
+                                    for (let i = 0; i < message.customSlackMessages.length; ++i) {
+                                        let error = $root.clutch.config.service.auditsink.slack.v1.CustomSlackMessage.verify(message.customSlackMessages[i]);
+                                        if (error)
+                                            return "customSlackMessages." + error;
+                                    }
+                                }
+                                return null;
+                            };
+
+                            /**
+                             * Creates an Override message from a plain object. Also converts values to their respective internal types.
+                             * @function fromObject
+                             * @memberof clutch.config.service.auditsink.slack.v1.Override
+                             * @static
+                             * @param {Object.<string,*>} object Plain object
+                             * @returns {clutch.config.service.auditsink.slack.v1.Override} Override
+                             */
+                            Override.fromObject = function fromObject(object) {
+                                if (object instanceof $root.clutch.config.service.auditsink.slack.v1.Override)
+                                    return object;
+                                let message = new $root.clutch.config.service.auditsink.slack.v1.Override();
+                                if (object.customSlackMessages) {
+                                    if (!Array.isArray(object.customSlackMessages))
+                                        throw TypeError(".clutch.config.service.auditsink.slack.v1.Override.customSlackMessages: array expected");
+                                    message.customSlackMessages = [];
+                                    for (let i = 0; i < object.customSlackMessages.length; ++i) {
+                                        if (typeof object.customSlackMessages[i] !== "object")
+                                            throw TypeError(".clutch.config.service.auditsink.slack.v1.Override.customSlackMessages: object expected");
+                                        message.customSlackMessages[i] = $root.clutch.config.service.auditsink.slack.v1.CustomSlackMessage.fromObject(object.customSlackMessages[i]);
+                                    }
+                                }
+                                return message;
+                            };
+
+                            /**
+                             * Creates a plain object from an Override message. Also converts values to other types if specified.
+                             * @function toObject
+                             * @memberof clutch.config.service.auditsink.slack.v1.Override
+                             * @static
+                             * @param {clutch.config.service.auditsink.slack.v1.Override} message Override
+                             * @param {$protobuf.IConversionOptions} [options] Conversion options
+                             * @returns {Object.<string,*>} Plain object
+                             */
+                            Override.toObject = function toObject(message, options) {
+                                if (!options)
+                                    options = {};
+                                let object = {};
+                                if (options.arrays || options.defaults)
+                                    object.customSlackMessages = [];
+                                if (message.customSlackMessages && message.customSlackMessages.length) {
+                                    object.customSlackMessages = [];
+                                    for (let j = 0; j < message.customSlackMessages.length; ++j)
+                                        object.customSlackMessages[j] = $root.clutch.config.service.auditsink.slack.v1.CustomSlackMessage.toObject(message.customSlackMessages[j], options);
+                                }
+                                return object;
+                            };
+
+                            /**
+                             * Converts this Override to JSON.
+                             * @function toJSON
+                             * @memberof clutch.config.service.auditsink.slack.v1.Override
+                             * @instance
+                             * @returns {Object.<string,*>} JSON object
+                             */
+                            Override.prototype.toJSON = function toJSON() {
+                                return this.constructor.toObject(this, $protobuf.util.toJSONOptions);
+                            };
+
+                            return Override;
                         })();
 
                         return v1;
@@ -14851,242 +15087,6 @@ export const clutch = $root.clutch = (() => {
                         };
 
                         return Filter;
-                    })();
-
-                    v1.CustomSlackMessage = (function() {
-
-                        /**
-                         * Properties of a CustomSlackMessage.
-                         * @memberof clutch.config.service.audit.v1
-                         * @interface ICustomSlackMessage
-                         * @property {string|null} [method] CustomSlackMessage method
-                         * @property {string|null} [message] CustomSlackMessage message
-                         */
-
-                        /**
-                         * Constructs a new CustomSlackMessage.
-                         * @memberof clutch.config.service.audit.v1
-                         * @classdesc Represents a CustomSlackMessage.
-                         * @implements ICustomSlackMessage
-                         * @constructor
-                         * @param {clutch.config.service.audit.v1.ICustomSlackMessage=} [properties] Properties to set
-                         */
-                        function CustomSlackMessage(properties) {
-                            if (properties)
-                                for (let keys = Object.keys(properties), i = 0; i < keys.length; ++i)
-                                    if (properties[keys[i]] != null)
-                                        this[keys[i]] = properties[keys[i]];
-                        }
-
-                        /**
-                         * CustomSlackMessage method.
-                         * @member {string} method
-                         * @memberof clutch.config.service.audit.v1.CustomSlackMessage
-                         * @instance
-                         */
-                        CustomSlackMessage.prototype.method = "";
-
-                        /**
-                         * CustomSlackMessage message.
-                         * @member {string} message
-                         * @memberof clutch.config.service.audit.v1.CustomSlackMessage
-                         * @instance
-                         */
-                        CustomSlackMessage.prototype.message = "";
-
-                        /**
-                         * Verifies a CustomSlackMessage message.
-                         * @function verify
-                         * @memberof clutch.config.service.audit.v1.CustomSlackMessage
-                         * @static
-                         * @param {Object.<string,*>} message Plain object to verify
-                         * @returns {string|null} `null` if valid, otherwise the reason why it is not
-                         */
-                        CustomSlackMessage.verify = function verify(message) {
-                            if (typeof message !== "object" || message === null)
-                                return "object expected";
-                            if (message.method != null && message.hasOwnProperty("method"))
-                                if (!$util.isString(message.method))
-                                    return "method: string expected";
-                            if (message.message != null && message.hasOwnProperty("message"))
-                                if (!$util.isString(message.message))
-                                    return "message: string expected";
-                            return null;
-                        };
-
-                        /**
-                         * Creates a CustomSlackMessage message from a plain object. Also converts values to their respective internal types.
-                         * @function fromObject
-                         * @memberof clutch.config.service.audit.v1.CustomSlackMessage
-                         * @static
-                         * @param {Object.<string,*>} object Plain object
-                         * @returns {clutch.config.service.audit.v1.CustomSlackMessage} CustomSlackMessage
-                         */
-                        CustomSlackMessage.fromObject = function fromObject(object) {
-                            if (object instanceof $root.clutch.config.service.audit.v1.CustomSlackMessage)
-                                return object;
-                            let message = new $root.clutch.config.service.audit.v1.CustomSlackMessage();
-                            if (object.method != null)
-                                message.method = String(object.method);
-                            if (object.message != null)
-                                message.message = String(object.message);
-                            return message;
-                        };
-
-                        /**
-                         * Creates a plain object from a CustomSlackMessage message. Also converts values to other types if specified.
-                         * @function toObject
-                         * @memberof clutch.config.service.audit.v1.CustomSlackMessage
-                         * @static
-                         * @param {clutch.config.service.audit.v1.CustomSlackMessage} message CustomSlackMessage
-                         * @param {$protobuf.IConversionOptions} [options] Conversion options
-                         * @returns {Object.<string,*>} Plain object
-                         */
-                        CustomSlackMessage.toObject = function toObject(message, options) {
-                            if (!options)
-                                options = {};
-                            let object = {};
-                            if (options.defaults) {
-                                object.method = "";
-                                object.message = "";
-                            }
-                            if (message.method != null && message.hasOwnProperty("method"))
-                                object.method = message.method;
-                            if (message.message != null && message.hasOwnProperty("message"))
-                                object.message = message.message;
-                            return object;
-                        };
-
-                        /**
-                         * Converts this CustomSlackMessage to JSON.
-                         * @function toJSON
-                         * @memberof clutch.config.service.audit.v1.CustomSlackMessage
-                         * @instance
-                         * @returns {Object.<string,*>} JSON object
-                         */
-                        CustomSlackMessage.prototype.toJSON = function toJSON() {
-                            return this.constructor.toObject(this, $protobuf.util.toJSONOptions);
-                        };
-
-                        return CustomSlackMessage;
-                    })();
-
-                    v1.Override = (function() {
-
-                        /**
-                         * Properties of an Override.
-                         * @memberof clutch.config.service.audit.v1
-                         * @interface IOverride
-                         * @property {Array.<clutch.config.service.audit.v1.ICustomSlackMessage>|null} [customSlackMessages] Override customSlackMessages
-                         */
-
-                        /**
-                         * Constructs a new Override.
-                         * @memberof clutch.config.service.audit.v1
-                         * @classdesc Represents an Override.
-                         * @implements IOverride
-                         * @constructor
-                         * @param {clutch.config.service.audit.v1.IOverride=} [properties] Properties to set
-                         */
-                        function Override(properties) {
-                            this.customSlackMessages = [];
-                            if (properties)
-                                for (let keys = Object.keys(properties), i = 0; i < keys.length; ++i)
-                                    if (properties[keys[i]] != null)
-                                        this[keys[i]] = properties[keys[i]];
-                        }
-
-                        /**
-                         * Override customSlackMessages.
-                         * @member {Array.<clutch.config.service.audit.v1.ICustomSlackMessage>} customSlackMessages
-                         * @memberof clutch.config.service.audit.v1.Override
-                         * @instance
-                         */
-                        Override.prototype.customSlackMessages = $util.emptyArray;
-
-                        /**
-                         * Verifies an Override message.
-                         * @function verify
-                         * @memberof clutch.config.service.audit.v1.Override
-                         * @static
-                         * @param {Object.<string,*>} message Plain object to verify
-                         * @returns {string|null} `null` if valid, otherwise the reason why it is not
-                         */
-                        Override.verify = function verify(message) {
-                            if (typeof message !== "object" || message === null)
-                                return "object expected";
-                            if (message.customSlackMessages != null && message.hasOwnProperty("customSlackMessages")) {
-                                if (!Array.isArray(message.customSlackMessages))
-                                    return "customSlackMessages: array expected";
-                                for (let i = 0; i < message.customSlackMessages.length; ++i) {
-                                    let error = $root.clutch.config.service.audit.v1.CustomSlackMessage.verify(message.customSlackMessages[i]);
-                                    if (error)
-                                        return "customSlackMessages." + error;
-                                }
-                            }
-                            return null;
-                        };
-
-                        /**
-                         * Creates an Override message from a plain object. Also converts values to their respective internal types.
-                         * @function fromObject
-                         * @memberof clutch.config.service.audit.v1.Override
-                         * @static
-                         * @param {Object.<string,*>} object Plain object
-                         * @returns {clutch.config.service.audit.v1.Override} Override
-                         */
-                        Override.fromObject = function fromObject(object) {
-                            if (object instanceof $root.clutch.config.service.audit.v1.Override)
-                                return object;
-                            let message = new $root.clutch.config.service.audit.v1.Override();
-                            if (object.customSlackMessages) {
-                                if (!Array.isArray(object.customSlackMessages))
-                                    throw TypeError(".clutch.config.service.audit.v1.Override.customSlackMessages: array expected");
-                                message.customSlackMessages = [];
-                                for (let i = 0; i < object.customSlackMessages.length; ++i) {
-                                    if (typeof object.customSlackMessages[i] !== "object")
-                                        throw TypeError(".clutch.config.service.audit.v1.Override.customSlackMessages: object expected");
-                                    message.customSlackMessages[i] = $root.clutch.config.service.audit.v1.CustomSlackMessage.fromObject(object.customSlackMessages[i]);
-                                }
-                            }
-                            return message;
-                        };
-
-                        /**
-                         * Creates a plain object from an Override message. Also converts values to other types if specified.
-                         * @function toObject
-                         * @memberof clutch.config.service.audit.v1.Override
-                         * @static
-                         * @param {clutch.config.service.audit.v1.Override} message Override
-                         * @param {$protobuf.IConversionOptions} [options] Conversion options
-                         * @returns {Object.<string,*>} Plain object
-                         */
-                        Override.toObject = function toObject(message, options) {
-                            if (!options)
-                                options = {};
-                            let object = {};
-                            if (options.arrays || options.defaults)
-                                object.customSlackMessages = [];
-                            if (message.customSlackMessages && message.customSlackMessages.length) {
-                                object.customSlackMessages = [];
-                                for (let j = 0; j < message.customSlackMessages.length; ++j)
-                                    object.customSlackMessages[j] = $root.clutch.config.service.audit.v1.CustomSlackMessage.toObject(message.customSlackMessages[j], options);
-                            }
-                            return object;
-                        };
-
-                        /**
-                         * Converts this Override to JSON.
-                         * @function toJSON
-                         * @memberof clutch.config.service.audit.v1.Override
-                         * @instance
-                         * @returns {Object.<string,*>} JSON object
-                         */
-                        Override.prototype.toJSON = function toJSON() {
-                            return this.constructor.toObject(this, $protobuf.util.toJSONOptions);
-                        };
-
-                        return Override;
                     })();
 
                     v1.SinkConfig = (function() {
